Break out functions into utils file
